pyproject.toml
[tool.poetry]
name = "demo-loyalty-api"
version = "0.1.0"
description = ""
authors = ["korchaginks <kkonstantinserg@gmail.com>"]
readme = "README.md"
packages = [{include = "demo_loyalty_api"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
django-phonenumber-field = "^7.3.0"
phonenumbers = "^8.13.35"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
pytest = "^8.1.1"
coverage = "^7.4.4"
psycopg2 = "^2.9.9"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
